:recycle: Pointer event patch has its own file
