+++ src/components/tabletop/Grid.spec.tsx
@@ -3,52 +3,14 @@ import { Provider } from "react-redux";
 import { combineReducers, createStore } from "redux";
 import { fireEvent } from "@testing-library/react";
 import { renderSVG } from "~/test-utilities";
+
 import "@testing-library/jest-dom/extend-expect";
+import "~/pointer-event";
 
 import Grid, { Grid as DisconnectedGrid, gridTestId } from "./Grid";
 import { viewTransform } from "~/reducers/tabletop";
 import { setViewTransform, SetViewTransformPayload } from "~/actions/tabletop";
 import { identityTransform, Transform } from "~/core/Transform";
-
-if (!global.PointerEvent)
-{
-  class PointerEvent extends MouseEvent
-  {
-    public pointerId?: number;
-    public pointerType?: string;
-
-    public isPrimary?: boolean;
-
-    public height?: number;
-    public width?: number;
-
-    public pressure?: number;
-    public tangentialPressure?: number;
-
-    public tiltX?: number;
-    public tiltY?: number;
-
-    public twist?: number;
-
-    constructor(type: string, params: PointerEventInit = {})
-    {
-      super(type, params);
-
-      this.pointerId = params.pointerId;
-      this.pointerType = params.pointerType;
-      this.isPrimary = params.isPrimary;
-      this.height = params.height;
-      this.width = params.width;
-      this.pressure = params.pressure;
-      this.tangentialPressure = params.tangentialPressure;
-      this.tiltX = params.tiltX;
-      this.tiltY = params.tiltY;
-      this.twist = params.twist;
-    }
-  }
-
-  global.PointerEvent = PointerEvent as any;
-}
 
 describe("Disconnected Grid component", () =>
 {
